feat: refactoring and adding code to pass unit test
