+++ src/AutoDrivingCarSimulator/Program.cs
@@ -1,6 +1,7 @@
 ﻿// See https://aka.ms/new-console-template for more information
 
 
+using AutoDrivingCarSimulator;
 using AutoDrivingCarSimulator.Core.DTO;
 using AutoDrivingCarSimulator.Core.Enums;
 using AutoDrivingCarSimulator.Core.Services;
@@ -11,153 +12,158 @@ var serviceProvider = new ServiceCollection()
     .ConfigureServices()
     .BuildServiceProvider();
 
-var simulationService = serviceProvider.GetRequiredService<ISimulatorService>();
+var app = serviceProvider.GetRequiredService<AutoDrivingCarApp>();
+app.Begin();
 
-Begin();
-void Begin()
-{
-    Console.WriteLine("Welcome to Auto Driving Car Simulation!");
-    bool isValidField = true;
-    int width = 0;  
-    int height = 0;
-    do
-    {
-        Console.WriteLine("Please enter the width and height of the simulation field in x y format:");
+/// Old Code
 
-        string[] coordinates = Console.ReadLine().Split(' ');
-        width = int.Parse(coordinates[0]);
-        height = int.Parse(coordinates[1]);
-        isValidField = simulationService.IsValidField(width, height);
+//var simulationService = serviceProvider.GetRequiredService<ISimulatorService>();
 
-        if (!isValidField)
-        {
-            Console.WriteLine("You have entered invalid width and height ");
-        }
-    } while (!isValidField);
+//Begin();
+//void Begin()
+//{
+//    Console.WriteLine("Welcome to Auto Driving Car Simulation!");
+//    bool isValidField = true;
+//    int width = 0;  
+//    int height = 0;
+//    do
+//    {
+//        Console.WriteLine("Please enter the width and height of the simulation field in x y format:");
 
-    // Initialize the simulation field
-    simulationService.AddField(width, height);
+//        string[] coordinates = Console.ReadLine().Split(' ');
+//        width = int.Parse(coordinates[0]);
+//        height = int.Parse(coordinates[1]);
+//        isValidField = simulationService.IsValidField(width, height);
 
-    while (true)
-    {
-        Console.WriteLine($"You have created a field of {width} x {height}.");
-        Console.WriteLine("Please choose from the following options:");
-        Console.WriteLine("[1] Add a car to field");
-        Console.WriteLine("[2] Run simulation");
+//        if (!isValidField)
+//        {
+//            Console.WriteLine("You have entered invalid width and height ");
+//        }
+//    } while (!isValidField);
 
-        var input = Console.ReadLine()?.ToUpper();
+//    // Initialize the simulation field
+//    simulationService.AddField(width, height);
 
-        switch (input)
-        {
-            case "1":
-                AddCar();
-                break;
-            case "2":
-                Simulate();
-                break;
-            case "Q" or "q":
-                Quit();
-                return;
-            default:
-                Console.WriteLine("Invalid option. Please try again.");
-                break;
-        }
-    }
-}
+//    while (true)
+//    {
+//        Console.WriteLine($"You have created a field of {width} x {height}.");
+//        Console.WriteLine("Please choose from the following options:");
+//        Console.WriteLine("[1] Add a car to field");
+//        Console.WriteLine("[2] Run simulation");
 
-void AddCar()
-{
-    bool isValidName = true;
-    string name = "";
-    bool isValidCar = true;
-    bool isValidCommand = true;
-    CarDto car;
+//        var input = Console.ReadLine()?.ToUpper();
 
-    do
-    {
-        Console.WriteLine("Please enter the name of the car:");
-        name = Console.ReadLine();
+//        switch (input)
+//        {
+//            case "1":
+//                AddCar();
+//                break;
+//            case "2":
+//                Simulate();
+//                break;
+//            case "Q" or "q":
+//                Quit();
+//                return;
+//            default:
+//                Console.WriteLine("Invalid option. Please try again.");
+//                break;
+//        }
+//    }
+//}
 
-        isValidName = simulationService.IsValidCarName(name);
+//void AddCar()
+//{
+//    bool isValidName = true;
+//    string name = "";
+//    bool isValidCar = true;
+//    bool isValidCommand = true;
+//    CarDto car;
 
-        if (!isValidName)
-        {
-            Console.WriteLine("You have entered an invalid name");
-        }
-    } while (!isValidName);
+//    do
+//    {
+//        Console.WriteLine("Please enter the name of the car:");
+//        name = Console.ReadLine();
 
-    do
-    {
-        Console.WriteLine($"Please enter initial position of car {name} in x y Direction format:");
-        string[] position = Console.ReadLine().Split(' ');
-        int x = int.Parse(position[0]);
-        int y = int.Parse(position[1]);
-        Direction direction = Enum.Parse<Direction>(position[2]);
+//        isValidName = simulationService.IsValidCarName(name);
 
-        car = new CarDto()
-        {
-            Name = name,
-            XCoordinate = x,
-            YCoordinate = y,
-            Direction = direction
-        };
+//        if (!isValidName)
+//        {
+//            Console.WriteLine("You have entered an invalid name");
+//        }
+//    } while (!isValidName);
 
-        isValidCar = simulationService.IsValidCar(car);
+//    do
+//    {
+//        Console.WriteLine($"Please enter initial position of car {name} in x y Direction format:");
+//        string[] position = Console.ReadLine().Split(' ');
+//        int x = int.Parse(position[0]);
+//        int y = int.Parse(position[1]);
+//        Direction direction = Enum.Parse<Direction>(position[2]);
 
-        if (!isValidCar)
-        {
-            Console.WriteLine("You have entered an invalid car details");
-        }
-    } while (!isValidCar);
+//        car = new CarDto()
+//        {
+//            Name = name,
+//            XCoordinate = x,
+//            YCoordinate = y,
+//            Direction = direction
+//        };
 
-    do
-    {
-        Console.WriteLine($"Please enter the commands for car {name}:");
-        string command = Console.ReadLine();
+//        isValidCar = simulationService.IsValidCar(car);
 
-        isValidCommand = simulationService.IsValidCommand(command.ToUpper());
+//        if (!isValidCar)
+//        {
+//            Console.WriteLine("You have entered an invalid car details");
+//        }
+//    } while (!isValidCar);
 
-        car.CommandList = command.ToUpper().Select(c => Enum.Parse<Command>(c.ToString())).ToList();
+//    do
+//    {
+//        Console.WriteLine($"Please enter the commands for car {name}:");
+//        string command = Console.ReadLine();
 
-        if (!isValidCommand)
-        {
-            Console.WriteLine("You have entered an invalid command");
-        }
+//        isValidCommand = simulationService.IsValidCommand(command.ToUpper());
 
-    } while (!isValidCommand);
+//        car.CommandList = command.ToUpper().Select(c => Enum.Parse<Command>(c.ToString())).ToList();
 
-    simulationService.AddCar(car);
+//        if (!isValidCommand)
+//        {
+//            Console.WriteLine("You have entered an invalid command");
+//        }
 
-    Console.WriteLine("Your current list of cars are:");
-    foreach (var c in simulationService.GetAllCars())
-    {
-        Console.WriteLine($"- {c.Name}, ({c.XCoordinate},{c.YCoordinate}) {c.Direction}, {string.Join("", c.CommandList)}");
-    }
-}
+//    } while (!isValidCommand);
 
-void Simulate()
-{
-    Console.WriteLine("Your current list of cars are:");
-    foreach (var c in simulationService.GetAllCars())
-    {
-        Console.WriteLine($"- {c.Name}, ({c.XCoordinate},{c.YCoordinate}) {c.Direction}, {string.Join("", c.CommandList)}");
-    }
+//    simulationService.AddCar(car);
 
-    simulationService.FindDestination();
+//    Console.WriteLine("Your current list of cars are:");
+//    foreach (var c in simulationService.GetAllCars())
+//    {
+//        Console.WriteLine($"- {c.Name}, ({c.XCoordinate},{c.YCoordinate}) {c.Direction}, {string.Join("", c.CommandList)}");
+//    }
+//}
 
-    Console.WriteLine("After simulation, the result is:");
-    foreach (var result in simulationService.GetResults())
-    {
-        Console.WriteLine(result);
-    }
+//void Simulate()
+//{
+//    Console.WriteLine("Your current list of cars are:");
+//    foreach (var c in simulationService.GetAllCars())
+//    {
+//        Console.WriteLine($"- {c.Name}, ({c.XCoordinate},{c.YCoordinate}) {c.Direction}, {string.Join("", c.CommandList)}");
+//    }
 
-    Console.WriteLine("Please choose from the following options:");
-    Console.WriteLine("[1] Start over");
-    Console.WriteLine("[2] Exit");
+//    simulationService.FindDestination();
 
-}
+//    Console.WriteLine("After simulation, the result is:");
+//    foreach (var result in simulationService.GetResults())
+//    {
+//        Console.WriteLine(result);
+//    }
 
-void Quit()
-{
-    Console.WriteLine("Thank you for running the simulation. Goodbye!");
-}
+//    Console.WriteLine("Please choose from the following options:");
+//    Console.WriteLine("[1] Start over");
+//    Console.WriteLine("[2] Exit");
+
+//}
+
+//void Quit()
+//{
+//    Console.WriteLine("Thank you for running the simulation. Goodbye!");
+//}
